+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.20
 - Add `sorbetto.maximumDiagnosticsCount` setting to control number of diagnostics reported by Sorbet.
+- Add `sorbetto.enableRubyfmt` setting to enable Rubyfmt integration in Sorbet.
 - Add `sorbetto.trace.server` setting for LSP communication logging.
 - Group all language server settings under a dedicated Sorbet section to improve clarity.